Move original-file upload out of pipeline/publish.rs
pipeline/original.rs holds upload_original_file_if_needed; publish.rs
drops to ~400 lines, completing the no-file-over-500-code-lines goal.
