+++ crates/rust_admin/src/pipeline.rs
@@ -1,9 +1,11 @@
 //! Video processing pipeline: transcode, quote, and publish stages.
 
 mod final_quote;
+mod original;
 mod process;
 mod publish;
 
 pub(crate) use final_quote::*;
+pub(crate) use original::*;
 pub(crate) use process::*;
 pub(crate) use publish::*;
